Rollback "Remove wiiupc", libwupc re-added.
Right stick wasn't working on Wii U Pro Controller.

+++ source/filebrowser.cpp
@@ -24,7 +24,6 @@
 #include <di/di.h>
 #endif
 
-#include "snes9x/port.h"
 #include "snes9xgx.h"
 #include "filebrowser.h"
 #include "menu.h"
@@ -455,7 +454,7 @@ int BrowserLoadSz()
 
 int WiiFileLoader()
 {
-	size_t size;
+	int size;
 	char filepath[1024];
 
 	memset(Memory.NSRTHeader, 0, sizeof(Memory.NSRTHeader));
